Refactor server into modules
